Don't remove git repos by default

+++ setup.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+force=0
+
 ENVDIR="${HOME}/.userenv"
 
 MYDIR="$(realpath "$(dirname "$0")")"
@@ -14,27 +16,57 @@ rm -f "$ENVDIR"
 ln -sf "$MYDIR" "$ENVDIR"
 cd "$ENVDIR"
 
-echo "Installing vimplug for neovim ..."
-mkdir -p ~/.local/share/nvim
-rm -f ~/.local/share/nvim/site/autoload/plug.vim
-curl -fLo ~/.local/share/nvim/site/autoload/plug.vim --create-dirs https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
+switch_git_fork() {
+	gitdir="$1"
+	repourl="$2"
+	branchname="$3"
+	remotename="$4"
+	cwd="`pwd`"
+	if ! echo "$repourl" | grep / >/dev/null; then
+		repourl="https://github.com/crispy1989/${repourl}.git"
+	fi
+	if [[ -z "$remotename" ]]; then
+		remotename=crispy
+	fi
+	cd "$gitdir"
+	if ! git remote | grep -F "$remotename" >/dev/null; then
+		echo "Switching out fork for $2 ..."
+		git remote add "$remotename" "$repourl"
+		git fetch "$remotename"
+		git checkout "$remotename"/"$branchname"
+	fi
+	cd "$cwd"
+}
 
-echo "Installing vundle for neovim ..."
-mkdir -p ~/.local/share/nvim/bundle
-rm -rf ~/.local/share/nvim/bundle/Vundle.vim
-git clone https://github.com/VundleVim/Vundle.vim.git ~/.local/share/nvim/bundle/Vundle.vim
-
-echo "Installing tpm for tmux ..."
-mkdir -p ~/.tmux/plugins/tpm
-git clone https://github.com/tmux-plugins/tpm ~/.tmux/plugins/tpm
-
-echo "Installing Oh My Zsh ..."
-export RUNZSH=no
-if [ -d ~/.oh-my-zsh ]; then
-	rm -rf ~/.oh-my-zsh-bak
-	mv ~/.oh-my-zsh ~/.oh-my-zsh-bak
+if [[ ! -e ~/.local/share/nvim/site/autoload/plug.vim ]] || [[ $force -eq 1 ]]; then
+	echo "Installing vimplug for neovim ..."
+	mkdir -p ~/.local/share/nvim
+	rm -f ~/.local/share/nvim/site/autoload/plug.vim
+	curl -fLo ~/.local/share/nvim/site/autoload/plug.vim --create-dirs https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
 fi
-sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
+
+if [[ ! -e ~/.local/share/nvim/bundle/Vundle.vim ]] || [[ $force -eq 1 ]]; then
+	echo "Installing vundle for neovim ..."
+	mkdir -p ~/.local/share/nvim/bundle
+	rm -rf ~/.local/share/nvim/bundle/Vundle.vim
+	git clone https://github.com/VundleVim/Vundle.vim.git ~/.local/share/nvim/bundle/Vundle.vim
+fi
+
+if [[ ! -e ~/.tmux/plugins/tpm ]] || [[ $force -eq 1 ]]; then
+	echo "Installing tpm for tmux ..."
+	mkdir -p ~/.tmux/plugins
+	git clone https://github.com/tmux-plugins/tpm ~/.tmux/plugins/tpm
+fi
+
+if [[ ! -e ~/.oh-my-zsh ]] || [[ $force -eq 1 ]]; then
+	echo "Installing Oh My Zsh ..."
+	export RUNZSH=no
+	if [ -d ~/.oh-my-zsh ]; then
+		rm -rf ~/.oh-my-zsh-bak
+		mv ~/.oh-my-zsh ~/.oh-my-zsh-bak
+	fi
+	sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
+fi
 
 echo "Setting up dotfiles ..."
 ./dotfiles/setup.sh
@@ -42,22 +74,15 @@ echo "Setting up dotfiles ..."
 echo "Setting up clipboard ..."
 ./clipboard/setup.sh
 
-echo "Installing fzf ..."
-rm -rf ~/.fzf
-git clone --depth 1 https://github.com/junegunn/fzf.git ~/.fzf
-~/.fzf/install --key-bindings --completion --no-update-rc
+if [[ ! -e ~/.fzf ]] || [[ $force -eq 1 ]]; then
+	echo "Installing fzf ..."
+	rm -rf ~/.fzf
+	git clone --depth 1 https://github.com/junegunn/fzf.git ~/.fzf
+	~/.fzf/install --key-bindings --completion --no-update-rc
+fi
 
 echo "Switching out personal forks ..."
-cd ~/.oh-my-zsh
-#git remote add crispy git@github.com:crispy1989/ohmyzsh.git
-git remote add crispy https://github.com/crispy1989/ohmyzsh.git &>/dev/null
-git fetch crispy
-git checkout crispy/master
-cd ~/.local/share/nvim/plugged/far.vim
-#git remote add crispy git@github.com:crispy1989/far.vim.git
-git remote add crispy https://github.com/crispy1989/far.vim.git &>/dev/null
-git fetch crispy
-git checkout crispy/master
-cd "$MYDIR"
+switch_git_fork ~/.oh-my-zsh ohmyzsh master
+switch_git_fork ~/.local/share/nvim/plugged/far.vim far.vim master
